Moved Complex to own controller
