+++ src/app/(Dashboard)/Landlord/Components/Grids/(Complex)/editComplexButton.tsx
@@ -3,7 +3,7 @@ import { EditIcon } from "../../Icons/EditIcon";
 import { Complex } from "../types";
 import { useRef, useState } from "react";
 import toast from "react-hot-toast";
-import { UpdateComplex } from "@/app/(Dashboard)/actions/landlordDashController";
+import { UpdateComplex } from "@/app/(Dashboard)/actions/landlordComplexController";
 
 
 // interface EditComplexbuttonProps {
